Utilizando paginação nas tabelas

+++ src/main/java/br/edu/ifpb/pweb2/plutao/repository/CursoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.ifpb.pweb2.plutao.model.Curso;
 import java.lang.Long;
 
-public interface CursoRepository extends JpaRepository<Curso, Long>{
+public interface CursoRepository extends JpaRepository<Curso, Integer>{
 
 }
